refactor: extract common sub-menu logic into reusable mixin
